feat: Add c-sharp, journaling and nicer looking prompt in eshell
I sometimes code in c-sharp and it is really good to have IntelliSense so I
added omnisharp for this. I also journal in Emacs so it is necessary to have a
journaling plugin and shortcuts for journaling in org mode. I upgrade eshell
prompt with a theme from eshell-git-prompt plugin since I use eshell so much.