feat: Modify org-refile default, agenda colours, bullets checkbox
I want to add a note when I refile in org-mode. Also Emacs normal agenda default
colours were not enough to help differentiate types of tasks at a glance so I
changed it. Finally, I want to be able to add check boxes to org bullets, so I
added that feature.