refactor: Change Yankpad's directory and refactor comments dividers
I decided to set my org-mode's directory and have Yankpad default to it. I think
this will be a better solution because I can now put my snippets in my Dropbox
folder or another repository which I don't have to share. This will now enable
me to share this repository without getting my personal snippets in the way if
others wants to use it. It is also a much cleaner solutions because I don't have
to commit and push to remote every time I capture a snippet.